chore: add test helpers module
